Spllit vers one adapter into helpers. Added serilog logging
