+++ ReadOtter/ReadOtter.csproj
@@ -87,6 +87,10 @@
       Version="$(MauiVersion)"
     />
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.2" />
+    <PackageReference Include="Serilog.Extensions.Logging" Version="8.0.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
+    <PackageReference Include="Serilog.Sinks.Debug" Version="3.0.0" />
+    <PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
     <PackageReference Include="VersOne.Epub" Version="3.3.2" />
   </ItemGroup>
   <ItemGroup>
